Store enums in separate files
