+++ .changeset/pink-mice-fly.md
@@ -1,0 +1,5 @@
+---
+'@tradeb0t/core': patch
+---
+
+Store enums in separate files
