+++ packages/core/src/domain/models/persistent/AlgorithmRunStatus.ts
@@ -1,0 +1,9 @@
+export enum EAlgorithmRunStatus {
+  RUNNING = 'running',
+  STOPPED = 'stopped',
+  RESUMED = 'resumed',
+  FINISHED = 'finished',
+  ERROR = 'error'
+}
+
+export type AlgorithmRunStatus = EAlgorithmRunStatus | `${EAlgorithmRunStatus}`
