+++ packages/core/src/domain/models/persistent/InputTypes.ts
@@ -1,0 +1,8 @@
+export enum EInputType {
+  STRING = 'string',
+  NUMBER = 'number',
+  ORDER_DETAILS = 'OrderDetails',
+  DATE = 'Date'
+}
+
+export type InputTypes = Record<string, EInputType | `${EInputType}`>
