+++ packages/core/src/domain/models/persistent/Order.ts
@@ -1,38 +1,8 @@
 import { Entity, Column, ManyToOne, UpdateDateColumn, PrimaryColumn } from 'typeorm'
 import { AlgorithmRun } from './AlgorithmRun'
 import { dateTransformer } from 'src/storage'
-
-/**
- * @see {@link https://support.tradeplusonline.com/support/solutions/articles/1000254592-what-are-the-different-order-status-possible-of-an-order- | What are the different order status possible of an order?}
- */
-export enum EOrderStatus {
-  NOT_PROCESSED = 'not_processed',
-  TO_BE_PROCESSED = 'to_be_processed',
-  PLACED = 'placed',
-  UNITS_ALLOCATED = 'units_allocated',
-  UNITS_REDEEMED = 'units_redeemed',
-  REJECTED = 'rejected',
-  CANCELLED = 'cancelled',
-  EXPIRED = 'expired',
-  UNDEFINED = 'undefined'
-}
-
-/**
- * @see {@link EOrderStatus}
- */
-export type OrderStatus = EOrderStatus | `${EOrderStatus}`
-
-export enum EOperationType {
-  LIMIT_BUY = 'limit_buy',
-  LIMIT_SELL = 'limit_sell',
-  MARKET_BUY = 'market_buy',
-  MARKET_SELL = 'market_sell',
-  BUY_OR_CANCEL = 'buy_or_cancel',
-  SELL_OR_CANCEL = 'sell_or_cancel',
-  UNDEFINED = 'undefined'
-}
-
-export type OperationType = EOperationType | `${EOperationType}`
+import { OrderStatus } from './OrderStatus'
+import { OperationType } from './OperationType'
 
 @Entity()
 export class Order {
