+++ packages/core/src/domain/models/persistent/index.ts
@@ -1,3 +1,7 @@
 export * from './AlgorithmRun'
 export * from './Algorithm'
 export * from './Order'
+export * from './OrderStatus'
+export * from './OperationType'
+export * from './InputTypes'
+export * from './AlgorithmRunStatus'
